Server/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
eth-account==0.11.0
eth-keys==0.5.0
eth-utils==2.3.1
Pillow
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
web3==6.11.1
python-dotenv==0.19.0
